Moved wave analysis code into the waves module directory. Use the pyjaia.waves module from both the adafruit BNO055 and BNO085 directories.
