+++ src/python/pyjaia/pyjaia/waves/series.py
