+++ src/python/pyjaia/pyjaia/waves/waves.md
