+++ src/python/pyjaia/pyproject.toml
@@ -1,0 +1,26 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pyjaia"
+dynamic = ["version"]
+description = "Jaia python utilities"
+license = ""
+authors = [
+    { name = "Jaia Robotics", email = "edsanville@gmail.com" },
+]
+dependencies = [
+    "cmocean",
+    "numpy",
+    "protobuf==3.20.0",
+    "scipy",
+    "turfpy",
+    "wheel",
+]
+
+[project.urls]
+Homepage = "https://www.jaia.tech"
+
+[project.scripts]
+jaia-analyze-waves = "pyjaia.waves.jaia-analyze-waves:main"
